Create strchr
The C++ Library strchr() function refers to the first occurrence of a character from a given string. This function operates on null characters (\0) or null-ended strings. This function is useful in a variety of text processing tasks where the user needs to locate a particular character.

Syntax
Following is the syntax of the C++ library strchr() function −

char *strchr(const char *str, int search_str);


Parameters
This function accepts the following parameters −

str − This parameter represents a given string.

search_str − This parameter refers to the specific character to be searched in a given string.

Return Value
This function returns a pointer to the first occurrence of the character i.e. search_str in the string (str), or NULL if the character is not found.